+++ test/framework/src/main/java/org/opensearch/test/InternalTestCluster.java
@@ -2721,6 +2721,7 @@ public final class InternalTestCluster extends TestCluster {
                     false,
                     false,
                     false,
+                    false,
                     false
                 );
                 assertThat(